Adding control when trying to delete teacher with student assignation

+++ src/main/java/org/sanmarcux/controller/impl/TeacherControllerImpl.java
@@ -25,7 +25,7 @@ public class TeacherControllerImpl implements TeacherController {
     public Object[][] getAll() {
         List<Teacher> teachers = dao.selectAll();
 
-        Object[][] rowData = new Object[teachers.size()][4];
+        Object[][] rowData = new Object[teachers.size()][5];
 
         for (int i = 0; i < teachers.size(); i++) {
             Teacher teacher = teachers.get(i);
@@ -34,12 +34,13 @@ public class TeacherControllerImpl implements TeacherController {
             rowData[i][1] = teacher.getNames();
             rowData[i][2] = dateFormatHelper.format(teacher.getBirthday());
             rowData[i][3] = teacher.getEmail();
+            rowData[i][4] = teacher.getAssignedStudents();
         }
         return rowData;
     }
 
     @Override
-    public Teacher getByCode(String code) {
+    public Teacher getByCode(final String code) {
         return dao.selectByCode(code);
     }
 
@@ -47,11 +48,11 @@ public class TeacherControllerImpl implements TeacherController {
     public Teacher[] getNames() {
         List<Teacher> teachers = dao.selectNames();
 
-        return teachers.toArray(new Teacher[teachers.size()]);
+        return teachers.toArray(new Teacher[0]);
     }
 
     @Override
-    public boolean saveOrUpdate(Teacher entity) {
+    public boolean saveOrUpdate(final Teacher entity) {
         int state;
 
         if (entity.getId() == 0) {
@@ -64,17 +65,22 @@ public class TeacherControllerImpl implements TeacherController {
     }
 
     @Override
-    public boolean delete(String code) {
-        Teacher entity = new Teacher();
-        entity.setCode(code);
+    public boolean delete(final String code) {
+        int result = dao.findAssignedStudents(code);
+        if (result == 0) {
+            Teacher entity = new Teacher();
+            entity.setCode(code);
 
-        int state = dao.delete(entity);
+            int state = dao.delete(entity);
 
-        return state != 0;
+            return state != 0;
+        } else {
+            return false;
+        }
     }
 
     @Override
-    public boolean existsCode(String code) {
+    public boolean existsCode(final String code) {
         return dao.selectIdByCode(code) != 0;
     }
